BackEnd/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
openai==1.45.0
python-dotenv==1.0.1
SQLAlchemy==2.0.34
Flask_Cors==5.0.0
python-dotenv==1.0.1
Requests==2.32.3
gunicorn==23.0.0
